feat: integrate RetryConfig into WebSocketConnector and update related tests
